requirements.txt
# Core dependencies
openai>=1.0.0
pillow>=9.0.0
requests>=2.28.0
tqdm>=4.64.0
scipy>=1.7.0

python-dotenv>=0.19.0
matplotlib==3.10.6
numpy==1.26.4
Pillow==10.4.0
qwen_vl_utils==0.0.14
transformers==4.57.3
accelerate==1.10.1
gradio==4.44.1
gradio_image_prompter==0.1.0
pydantic==2.10.6
triton
vllm==0.15.0
git+https://github.com/huggingface/diffusers
